[Refactor] 사용자 프로필 부분 Profile 컴포넌트로 대체
